src/requirements.txt
# set a fixed version of python-telegram-bot due to constantly appearing bugs: https://github.com/python-telegram-bot/python-telegram-bot/issues/4181
python-telegram-bot==21.1.1
SQLAlchemy==2.0.23
boto3==1.28.78
botocore==1.31.78
psycopg2-binary==2.9.9
Pillow==10.1.0
httpx==0.27.0